Polished Blog Command
Made the changes requested from the original pull request. Moved blog function to dotabase, swapped to the httpgetter, removed async io from the requiremnets and deleted extraneous functions.
