+++ requirements.txt
@@ -14,4 +14,4 @@ databases==0.2.2
 aiosqlite==0.10.0
 discord.py[voice]==1.3.3
 feedparser >= 6.0.2
-asyncio >= 3.4.3
+
